requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# Cloud platform APIs
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
kaggle==1.6.14

# Machine Learning frameworks (optional)
torch==2.1.0
tensorflow==2.15.0
transformers==4.36.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Utilities
tqdm==4.67.0
PyYAML==6.0.1
python-dateutil==2.8.2

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
